Extract targets parser and compat data from preset-env
