+++ packages/babel-compat-data/.npmignore
@@ -1,0 +1,4 @@
+scripts
+src
+test
+*.log
