+++ packages/babel-compat-data/corejs2-built-ins.js
@@ -1,0 +1,4 @@
+// Node < 13.3 doesn't support export maps in package.json.
+// Use this proxy file as a fallback.
+
+module.exports = require("./data/corejs2-built-ins.json");
