+++ packages/babel-compat-data/data/built-in-modules.json
@@ -1,0 +1,12 @@
+{
+  "es6.module": {
+    "edge": "16",
+    "firefox": "60",
+    "chrome": "61",
+    "safari": "10.1",
+    "opera": "48",
+    "ios_saf": "10.3",
+    "and_chr": "74",
+    "and_ff": "66"
+  }
+}
