+++ packages/babel-compat-data/data/native-modules.json
