+++ packages/babel-compat-data/data/overlapping-plugins.json
@@ -1,0 +1,3 @@
+{
+  "transform-regenerator": []
+}
