+++ packages/babel-compat-data/native-modules.js
@@ -1,0 +1,4 @@
+// Node < 13.3 doesn't support export maps in package.json.
+// Use this proxy file as a fallback.
+
+module.exports = require("./data/native-modules.json");
