+++ packages/babel-compat-data/plugins.js
@@ -1,0 +1,4 @@
+// Node < 13.3 doesn't support export maps in package.json.
+// Use this proxy file as a fallback.
+
+module.exports = require("./data/plugins.json");
