+++ packages/babel-compat-data/scripts/build-modules-support.js
