+++ packages/babel-compat-data/scripts/data/overlapping-plugins.js
@@ -4,13 +4,13 @@ module.exports = new Map();
 
 // async -> regenerator is better than async -> generator -> regenerator
 ifIncluded("transform-regenerator")
-  // https://github.com/babel/babel/issues/10678
+  // Temporarly disabled: https://github.com/babel/babel/issues/10678
   // .isUnnecessary("transform-async-to-generator");
 
 function ifIncluded(name) {
   const set = new Set();
   module.exports.set(name, set);
   return {
-    isUnnecessary(name) { set.add(name); return this; }
+    isUnnecessary(name) { set.add(name); return this; },
   };
 }
