+++ packages/babel-helper-compilation-targets/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+*.log
