+++ packages/babel-helper-compilation-targets/src/options.js
@@ -1,0 +1,20 @@
+// @flow
+
+export const TargetNames = {
+  esmodules: "esmodules",
+  node: "node",
+  browsers: "browsers",
+  chrome: "chrome",
+  opera: "opera",
+  edge: "edge",
+  firefox: "firefox",
+  safari: "safari",
+  ie: "ie",
+  ios: "ios",
+  android: "android",
+  electron: "electron",
+  samsung: "samsung",
+
+  // deprecated
+  uglify: "uglify",
+};
