+++ packages/babel-helper-compilation-targets/src/targets.js
@@ -1,0 +1,20 @@
+export const unreleasedLabels = {
+  safari: "tp",
+};
+
+export const browserNameMap = {
+  and_chr: "chrome",
+  and_ff: "firefox",
+  android: "android",
+  chrome: "chrome",
+  edge: "edge",
+  firefox: "firefox",
+  ie: "ie",
+  ie_mob: "ie",
+  ios_saf: "ios",
+  node: "node",
+  op_mob: "opera",
+  opera: "opera",
+  safari: "safari",
+  samsung: "samsung",
+};
