+++ packages/babel-helper-compilation-targets/src/types.js
@@ -1,0 +1,19 @@
+// @flow
+
+// Targets
+export type Target =
+  | "node"
+  | "chrome"
+  | "opera"
+  | "edge"
+  | "firefox"
+  | "safari"
+  | "ie"
+  | "ios"
+  | "android"
+  | "electron"
+  | "samsung";
+
+export type Targets = {
+  [target: Target]: string,
+};
