+++ packages/babel-helper-compilation-targets/test/pretty.spec.js
@@ -1,27 +1,6 @@
-"use strict";
+import { prettifyTargets, prettifyVersion } from "../lib/pretty";
 
-const utils = require("../lib/utils");
-
-const { prettifyTargets, prettifyVersion, semverify } = utils;
-
-describe("utils", () => {
-  describe("semverify", () => {
-    it("returns", () => {
-      expect(semverify("1")).toBe("1.0.0");
-      expect(semverify("1.0")).toBe("1.0.0");
-      expect(semverify("1.0.0")).toBe("1.0.0");
-      expect(semverify(1)).toBe("1.0.0");
-      expect(semverify(1.2)).toBe("1.2.0");
-    });
-
-    it("throws", () => {
-      const invalidSemver = () => {
-        semverify("invalid");
-      };
-      expect(invalidSemver).toThrow();
-    });
-  });
-
+describe("pretty", () => {
   describe("prettifyVersion", () => {
     it("returns", () => {
       expect(prettifyVersion(true)).toBe(true);
