+++ packages/babel-preset-env/data/built-ins.js
@@ -1,0 +1,4 @@
+// TODO: Remove in Babel 8
+// https://github.com/vuejs/vue-cli/issues/3671
+
+module.exports = require("./corejs2-built-ins.json");
