+++ packages/babel-preset-env/data/corejs2-built-ins.json.js
@@ -1,0 +1,3 @@
+// TODO: Remove in Babel 8
+
+module.exports = require("@babel/compat-data/corejs2-built-ins");
