+++ packages/babel-preset-env/data/unreleased-labels.js
@@ -1,3 +1,3 @@
-module.exports = {
-  safari: "tp",
-};
+// TODO: Remove in Babel 8
+
+module.exports = require("@babel/helper-compilation-targets").unreleasedLabels;
