+++ packages/babel-preset-env/package.json
@@ -10,10 +10,9 @@
   },
   "repository": "https://github.com/babel/babel/tree/master/packages/babel-preset-env",
   "main": "lib/index.js",
-  "scripts": {
-    "build-data": "./scripts/download-compat-table.sh; node ./scripts/build-data.js; node ./scripts/build-modules-support.js"
-  },
   "dependencies": {
+    "@babel/compat-data": "^0.0.0",
+    "@babel/helper-compilation-targets": "^0.0.0",
     "@babel/helper-module-imports": "^7.7.4",
     "@babel/helper-plugin-utils": "^7.0.0",
     "@babel/plugin-proposal-async-generator-functions": "^7.7.4",
@@ -74,8 +73,6 @@
     "@babel/core": "^7.7.7",
     "@babel/helper-fixtures": "^7.6.3",
     "@babel/helper-plugin-test-runner": "^7.7.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "caniuse-db": "1.0.30000969",
-    "electron-to-chromium": "1.3.113"
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0"
   }
 }
