+++ packages/babel-preset-env/src/options.js
@@ -31,22 +31,3 @@ export const UseBuiltInsOption = {
   entry: "entry",
   usage: "usage",
 };
-
-export const TargetNames = {
-  esmodules: "esmodules",
-  node: "node",
-  browsers: "browsers",
-  chrome: "chrome",
-  opera: "opera",
-  edge: "edge",
-  firefox: "firefox",
-  safari: "safari",
-  ie: "ie",
-  ios: "ios",
-  android: "android",
-  electron: "electron",
-  samsung: "samsung",
-
-  // deprecated
-  uglify: "uglify",
-};
