+++ packages/babel-preset-env/src/polyfills/corejs2/usage-plugin.js
@@ -1,8 +1,8 @@
 // @flow
 
-import corejs2Polyfills from "../../../data/corejs2-built-ins.json";
+import corejs2Polyfills from "@babel/compat-data/corejs2-built-ins";
+import { filterItems } from "@babel/helper-compilation-targets";
 import getPlatformSpecificDefaultFor from "./get-platform-specific-default";
-import filterItems from "../../filter-items";
 import {
   BuiltIns,
   StaticProperties,
