+++ packages/babel-preset-env/src/utils.js
@@ -2,12 +2,8 @@
 
 import * as t from "@babel/types";
 import type { NodePath } from "@babel/traverse";
-import invariant from "invariant";
-import semver from "semver";
 import { addSideEffect } from "@babel/helper-module-imports";
-import unreleasedLabels from "../data/unreleased-labels";
-import { semverMin } from "./targets-parser";
-import type { Target, Targets } from "./types";
+import type { Targets } from "@babel/helper-compilation-targets";
 
 export const has = Object.hasOwnProperty.call.bind(Object.hasOwnProperty);
 
@@ -18,28 +14,6 @@ export function getType(target: any): string {
     .toLowerCase();
 }
 
-const versionRegExp = /^(\d+|\d+.\d+)$/;
-
-// Convert version to a semver value.
-// 2.5 -> 2.5.0; 1 -> 1.0.0;
-export function semverify(version: number | string): string {
-  if (typeof version === "string" && semver.valid(version)) {
-    return version;
-  }
-
-  invariant(
-    typeof version === "number" ||
-      (typeof version === "string" && versionRegExp.test(version)),
-    `'${version}' is not a valid version`,
-  );
-
-  const split = version.toString().split(".");
-  while (split.length < 3) {
-    split.push("0");
-  }
-  return split.join(".");
-}
-
 export function intersection<T>(
   first: Set<T>,
   second: Set<T>,
@@ -48,71 +22,6 @@ export function intersection<T>(
   const result = new Set();
   for (const el of first) {
     if (second.has(el) && third.has(el)) result.add(el);
-  }
-  return result;
-}
-
-export function prettifyVersion(version: string) {
-  if (typeof version !== "string") {
-    return version;
-  }
-
-  const parts = [semver.major(version)];
-  const minor = semver.minor(version);
-  const patch = semver.patch(version);
-
-  if (minor || patch) {
-    parts.push(minor);
-  }
-
-  if (patch) {
-    parts.push(patch);
-  }
-
-  return parts.join(".");
-}
-
-export function prettifyTargets(targets: Targets): Targets {
-  return Object.keys(targets).reduce((results, target) => {
-    let value = targets[target];
-
-    const unreleasedLabel = unreleasedLabels[target];
-    if (typeof value === "string" && unreleasedLabel !== value) {
-      value = prettifyVersion(value);
-    }
-
-    results[target] = value;
-    return results;
-  }, {});
-}
-
-export function isUnreleasedVersion(
-  version: string | number,
-  env: string,
-): boolean {
-  const unreleasedLabel = unreleasedLabels[env];
-  return (
-    !!unreleasedLabel && unreleasedLabel === version.toString().toLowerCase()
-  );
-}
-
-export function getLowestUnreleased(a: string, b: string, env: string): string {
-  const unreleasedLabel = unreleasedLabels[env];
-  const hasUnreleased = [a, b].some(item => item === unreleasedLabel);
-  if (hasUnreleased) {
-    return a === hasUnreleased ? b : a || b;
-  }
-  return semverMin(a, b);
-}
-
-export function getLowestImplementedVersion(
-  plugin: Targets,
-  environment: Target,
-): string {
-  const result = plugin[environment];
-  // When Android support data is absent, use Chrome data as fallback
-  if (!result && environment === "android") {
-    return plugin.chrome;
   }
   return result;
 }
